backend/requirements.txt
asposestorage==1.0.2
fastapi==0.115.12
motor==3.7.0
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
pydantic==2.11.3
pymongo==4.12.0
python-dotenv==1.1.0
python_jose==3.4.0
python-multipart==0.0.20
Requests==2.32.3
routers==0.10.1
uvicorn==0.34.0
xlsxwriter==3.2.2
